+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Pass `self` directly to `Cycle.wrapped.iter` in `Iter.cycle`
+- Struct marking now uses a generic type `Marker`
 
 ## [0.5.0] - 2025-06-02